requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==1.10.15
redis==5.0.8
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
requests==2.32.3
pylint==3.3.0
fakeredis==2.31.3